make MAX_OCTAVES int

+++ src/Components/StepStrip.cpp
@@ -48,7 +48,7 @@ StepStrip::StepStrip(State &s, stepnum_t n) : stepNum(n), state(s) {
         octaveLabel.attachToComponent(&octaveMenu, false);
     }
     int itemId = 1;
-    for (int i = static_cast<int>(constants::MAX_OCTAVES); i >= -static_cast<int>(constants::MAX_OCTAVES); i--)
+    for (int i = constants::MAX_OCTAVES; i >= -constants::MAX_OCTAVES; i--)
         octaveMenu.addItem(std::to_string(i), itemId++);
     addAndMakeVisible(octaveMenu);
     octaveAttachment = std::make_unique<ComboBoxAttachment>(
