+++ src/Constants.h
@@ -23,7 +23,7 @@ namespace constants {
     constexpr size_t MAX_NOTES = 24;
     constexpr stepnum_t MAX_STEPS = 16;
     constexpr voicenum_t MAX_VOICES = 8;
-    constexpr size_t MAX_OCTAVES = 4;
+    constexpr int MAX_OCTAVES = 4;
 
     const juce::Colour COLOR_STANDARD = juce::Colours::white;
     const juce::Colour COLOR_SLIDER = juce::Colours::white;
